Config update student form and move from class to function into show update studen component
