--- front-end/student-react/archive/CRUD/CRUD.css
@@ -1,12 +1,0 @@
-table,th,td {
-  border: 1px solid;
-}
-
-table {
-  width: 100%;
-}
-
-button.space{
-    margin-right: 5%;
-    text-align: center;
-}
