+++ front-end/student-react/src/App.js
@@ -3,7 +3,6 @@ import "./App.css";
 import "bootstrap";
 import Student from "./components/Student/Student.jsx";
 import NavigationBar from "./components/NavigationBar/NavigationBar";
-import CRUD from "./components/CRUD/CRUD";
 
 
 export function App() {
